requirements.txt
catkin_pkg
rospy
openai
langchain
faiss-cpu
langchain_community
langchain-openai
python-dotenv
